requirements.txt
flask
requests
google-generativeai
python-dotenv
tiktoken 
werkzeug